+++ example/stairs/round/greebled_stairs.py
@@ -1,0 +1,15 @@
+import cadquery as cq
+from cqterrain.stairs.round import greebled_stairs
+
+stairs = greebled_stairs(
+    stair_count = 12,
+    height = 75,
+    inner_diameter = 75,
+    diameter = 75 + 55, 
+    debug = False
+).translate((0,0,(75)/2))
+
+#show_object(stairs)
+
+cq.exporters.export(stairs,'stl/stairs_round_greebled_stairs.stl')
+